output/services/legal-service/pyproject.toml
[project]
name = "venturestrat-legal-service"
version = "1.0.0"
description = "Venturestrat legal-service"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "VentureStrat Team"}
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=24.1.0",
    "httpx>=0.26.0",


    # Database (required for BaseModel ORM)
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.25",

    # Legal document generation
    "jinja2>=3.1.0",
    "python-docx>=1.1.0",
    "markdown>=3.5.0",

    # Platform integration
    "python-consul>=1.1.0",
    "aiokafka>=0.10.0",

    # Observability
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.22.0",

    # Platform SDKs
#     "venturestrat-registry @ file:///app/services/registry-service/sdk",  # Commented: works only inside Docker
#     "venturestrat-schema-registry @ file:///app/packages/schema-registry",  # Commented: works only inside Docker
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "testcontainers>=3.7.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/legal_service"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (require Docker)",
    "slow: Slow tests (>5 seconds)",
]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
